src/main/java/com/es/phoneshop/web/OrderOverviewPageServlet.java
package com.es.phoneshop.web;

import com.es.phoneshop.order.Order;
import com.es.phoneshop.order.OrderDao;
import com.es.phoneshop.order.OrderDaoImpl;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.util.Optional;

public class OrderOverviewPageServlet extends HttpServlet {
    private OrderDao orderDao;

    @Override
    public void init() throws ServletException {
        orderDao = OrderDaoImpl.getInstance();
    }

    @Override
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        String secureId = request.getPathInfo().substring(1);
        Optional<Order> optionalOrder = orderDao.getBySecureId(secureId);
        request.setAttribute("order", optionalOrder.get());
        request.getRequestDispatcher("/WEB-INF/pages/orderOverview.jsp").forward(request, response);
    }
}
